requirements.txt
pandas
numpy
streamlit
yfinance
PyYAML
akshare
pdfplumber

# OCR相关依赖
easyocr
opencv-python
Pillow
plotly
